+++ src/tools/node_maintain.py
@@ -1,18 +1,13 @@
-import requests
 import logging
-import os
-import sys
 import argparse
 import time
-import subprocess
 import re
 
-sys.path.append("../..")
-from deployment.paiLibrary.common.kubernetes_handler import get_configmap, update_configmap
-from deployment.k8sPaiLibrary.maintainlib import common
+from operator_wrapper import AlertOperator, KubernetesOperator, YarnOperator
 
 
 def setup_logger():
+    global logger
     logger = logging.getLogger(__name__)
     logger.setLevel(logging.DEBUG)
     ch = logging.StreamHandler()
@@ -21,229 +16,6 @@ def setup_logger():
     ch.setFormatter(formatter)
     logger.addHandler(ch)
     return logger
-
-
-class KubernetesOperator(object):
-    kubernetes_template = "../../deployment/k8sPaiLibrary/template/config.template"
-    kube_config_path = "./.config"
-    configmap_name = "exclude-file"
-    configmap_data_key = "nodes"
-
-    def __init__(self, master_ip):
-        self.master_ip = master_ip
-        self.setup_kubernetes_configfile(master_ip)
-
-    def setup_kubernetes_configfile(self, api_servers_ip):
-
-        template_data = common.read_template(self.kubernetes_template)
-        dict_map = {
-            "cluster_cfg": {"kubernetes": {"api-servers-ip": api_servers_ip}},
-        }
-        generated_data = common.generate_from_template_dict(template_data, dict_map)
-
-        common.write_generated_file(generated_data, self.kube_config_path)
-
-    def get_nodes(self):
-        configmap_info = get_configmap(self.kube_config_path, self.configmap_name)
-        nodes_str = configmap_info["data"][self.configmap_data_key]
-        nodes = set(nodes_str.splitlines())
-        return nodes
-
-    def set_nodes(self, nodes):
-        nodes = set(nodes)
-        nodes_str = '\n'.join(nodes)
-        data_dict = {self.configmap_data_key: nodes_str}
-        update_configmap(self.kube_config_path, self.configmap_name, data_dict)
-
-
-class YarnOperator(object):
-    yarn_config_path = "./.hadoop"
-
-    def __init__(self, master_ip):
-        self.master_ip = master_ip
-        self.setup_yarn_configfile(master_ip)
-
-    def setup_yarn_configfile(self, yarn_ip):
-        if not os.path.exists(self.yarn_config_path):
-            os.mkdir(self.yarn_config_path)
-
-        yarn_config_str = \
-        '''<configuration>
-            <property>
-                <name>yarn.resourcemanager.hostname</name>
-                <value>{}</value>
-            </property>
-        </configuration>'''.format(yarn_ip)
-
-        with open(os.path.join(self.yarn_config_path, "yarn-site.xml"), 'w') as f:
-            f.write(yarn_config_str)
-
-    def get_node_status(self):
-        yarn_nodes_url = "http://{}:8088/ws/v1/cluster/nodes".format(self.master_ip)
-        try:
-            response = requests.get(yarn_nodes_url, timeout=10)
-            response.raise_for_status()
-        except requests.exceptions.RequestException as e:
-            logger.exception(e)
-            sys.exit(1)
-
-        nodes_info = response.json()
-        current_nodes = {}
-        for node in nodes_info["nodes"]["node"]:
-            host, state = node["nodeHostName"], node["state"]
-            current_nodes[host] = state
-        return current_nodes
-
-    def decommission_nodes(self):
-        command = "yarn --config {} rmadmin -refreshNodes -g -server".format(self.yarn_config_path)
-        try:
-            subprocess.check_output(command, stderr=subprocess.STDOUT, shell=True)
-        except subprocess.CalledProcessError as e:
-            logger.error(e.output)
-            sys.exit(1)
-
-    def get_cluster_label(self):
-        # Sample output: "Node Labels: <label_ex:exclusivity=true>,<label_non:exclusivity=false>"
-        # Sample output: "Node Labels: "
-        command = "yarn --config {} cluster --list-node-labels".format(self.yarn_config_path)
-
-        try:
-            command_output = subprocess.check_output(command, shell=True).decode("utf8")
-        except subprocess.CalledProcessError as e:
-            logger.error(e.output)
-            return {}
-
-        lines = command_output.split("\n")
-        labels = dict()  # key: label name, value: exclusivity
-        for line in lines:
-            if not line.startswith("Node Labels:"):
-                continue
-            line = line.lstrip("Node Labels:")
-            labels_str = line.split(",")
-            label_regex = r"<([a-zA-Z0-9][a-zA-Z0-9_\-]*):exclusivity=(true|false)>"
-            for label_str in labels_str:
-                match = re.search(label_regex, label_str)
-                if match:
-                    label_name, exclusivity = match.groups()
-                    exclusivity = exclusivity == "true"
-                    labels[label_name] = exclusivity
-
-        return labels
-
-    def add_cluster_label(self, labels):
-        labels_list = []
-
-        for label_name in labels:
-            exclusivity = "true" if labels[label_name] else "false"
-            label_str = "{}(exclusive={})".format(label_name, exclusivity)
-            labels_list.append(label_str)
-
-        labels_str = ",".join(labels_list)
-
-        command = "yarn --config {} rmadmin -addToClusterNodeLabels \"{}\"".format(self.yarn_config_path, labels_str)
-        try:
-            subprocess.check_output(command, stderr=subprocess.STDOUT, shell=True).decode("utf8")
-        except subprocess.CalledProcessError as e:
-            logger.error(e.output)
-            sys.exit(1)
-
-    def remove_cluster_label(self, labels):
-        labels_str = ",".join(labels)   # Labels could be list, set or dict
-
-        command = "yarn --config {} rmadmin -removeFromClusterNodeLabels {}".format(self.yarn_config_path, labels_str)
-        try:
-            subprocess.check_output(command, stderr=subprocess.STDOUT, shell=True).decode("utf8")
-        except subprocess.CalledProcessError as e:
-            logger.error(e.output)
-            sys.exit(1)
-
-    def label_nodes(self, nodes):
-        nodes_list = []
-
-        for node in nodes:
-            node_str = "{}={}".format(node, nodes[node])
-            nodes_list.append(node_str)
-
-        nodes_str = " ".join(nodes_list)
-
-        # yarn rmadmin -replaceLabelsOnNode "node1[:port]=label1 node2=label2" [-failOnUnknownNodes]
-        command = "yarn --config {} rmadmin -replaceLabelsOnNode \"{}\" -failOnUnknownNodes"\
-            .format(self.yarn_config_path, nodes_str)
-        try:
-            subprocess.check_output(command, stderr=subprocess.STDOUT, shell=True).decode("utf8")
-        except subprocess.CalledProcessError as e:
-            logger.error(e.output)
-            sys.exit(1)
-
-    def unlabel_nodes(self, nodes):
-        nodes_list = []
-
-        for node in nodes:
-            node_str = "{}=".format(node)
-            nodes_list.append(node_str)
-
-        nodes_str = " ".join(nodes_list)
-
-        command = "yarn --config {} rmadmin -replaceLabelsOnNode \"{}\" -failOnUnknownNodes" \
-            .format(self.yarn_config_path, nodes_str)
-        try:
-            subprocess.check_output(command, stderr=subprocess.STDOUT, shell=True).decode("utf8")
-        except subprocess.CalledProcessError as e:
-            logger.error(e.output)
-            sys.exit(1)
-
-    def get_node_label(self):
-        yarn_nodes_url = "http://{}:8088/ws/v1/cluster/nodes".format(self.master_ip)
-        try:
-            response = requests.get(yarn_nodes_url, timeout=10)
-            response.raise_for_status()
-        except requests.exceptions.RequestException as e:
-            logger.exception(e)
-            sys.exit(1)
-
-        nodes_info = response.json()
-        current_nodes = {}
-        for node in nodes_info["nodes"]["node"]:
-            host = node["nodeHostName"]
-            node_label = node.get("nodeLabels", "")
-            current_nodes[host] = node_label
-        return current_nodes
-
-    def _generate_queue_xml(self, g):
-        pass
-
-
-class AlertOperator(object):
-    ALERT_TYPE = {
-        "gpu_related": {"NvidiaSmiLatencyTooLarge", "NvidiaSmiEccError", "NvidiaMemoryLeak", "NvidiaZombieProcess", "GpuUsedByExternalProcess", "GpuUsedByZombieContainer"},
-    }
-
-    def __init__(self, prometheus_ip, prometheus_port=9091):
-        self.master_ip = prometheus_ip
-        self.master_port = prometheus_port
-        self.alert_manager_url = "http://{}:{}/prometheus/api/v1/query?query=ALERTS".format(prometheus_ip, prometheus_port)
-
-    def get_gpu_alert_nodes(self):
-        try:
-            response = requests.get(self.alert_manager_url, timeout=10)
-            response.raise_for_status()
-        except requests.exceptions.RequestException as e:
-            logger.exception(e)
-            sys.exit(1)
-
-        if response.json()["status"] != "success":
-            logger.error("Alert response error: {}".format(response.text))
-            sys.exit(1)
-
-        alerts_info = response.json()["data"]["result"]
-        gpu_alert_nodes = {}
-        for alert in alerts_info:
-            metric = alert["metric"]
-            if metric["alertname"] in self.ALERT_TYPE["gpu_related"] and metric["alertstate"] == "firing":
-                node_ip = metric["instance"].split(':')[0]
-                gpu_alert_nodes[node_ip] = metric["alertname"]
-
-        return gpu_alert_nodes
 
 
 def get_unready_nodes(decommissioned_nodes, current_status):
