+++ src/tools/tests/test_node_maintain.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import
+
+from mock import patch
+import unittest
+
+from node_maintain import *
+
+
+class NodeMaintainTestCase(unittest.TestCase):
+
+    def setUp(self):
+        pass
+
+
+if __name__ == "__main__":
+    unittest.main()
